Finish the migration away from bash for all but testing.
Well, Ubuntu still has bash by default but Alpine doesn't and we don't add it
for anything but the :bash tag.  We don't use or need bash for anything but the
testing really, and debugging, but default sh on Alpine is alright compared to
default sh prompt on Ubuntu so that's not a big deal.